requirements.txt
addict==2.3.0
albumentations==0.5.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mmcv==1.2.0
mmcv-full==1.1.5
mmdet==2.4.0
mmpycocotools==12.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.4
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycocotools==2.0.2
pycparser==2.20
Pygments==2.7.2
pyparsing==3.0.0b1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
tifffile==2020.9.3
torch==1.5.0
torchvision==0.6.0
tornado==6.1
tqdm==4.52.0
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yapf==0.30.0
zipp==3.4.0
